# shared/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: ChatMod Mobile API
  version: 0.1.0
  description: Backend contract for the phone-hosted YouTube Live moderation bot.
servers:
  - url: http://localhost:4100
paths:
...
  /overlays/profiles/{profileId}:
    get:
      summary: Get OBS/browser overlay configuration for a channel profile
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: profileId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Overlay configuration and entitlement status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OverlayConfig"
    put:
      summary: Create or update a Pro-gated OBS/browser overlay configuration
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: profileId
          required: true
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OverlayConfigUpdate"
      responses:
        "200":
          description: Overlay configuration saved; public URL is returned only when a token is created or rotated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OverlayConfigMutation"
        "403":
          description: OBS/browser overlay requires Pro or Creator plan
  /overlays/profiles/{profileId}/rotate-token:
    post:
      summary: Rotate the read-only public OBS overlay token
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: profileId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Overlay token rotated and a fresh public URL returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OverlayConfigMutation"
        "403":
          description: OBS/browser overlay requires Pro or Creator plan
  /overlays/public/{token}:
    get:
      summary: Public read-only HTML overlay for OBS browser sources
      parameters:
        - in: path
          name: token
          required: true
          schema:
            type: string
            pattern: "^cmo_[A-Za-z0-9_-]{32,128}$"
      responses:
        "200":
          description: Transparent-friendly overlay HTML
          content:
            text/html:
              schema:
                type: string
        "404":
          description: Overlay token not found
  /overlays/public/{token}/state:
    get:
      summary: Public read-only overlay state for OBS polling
      parameters:
        - in: path
          name: token
          required: true
          schema:
            type: string
            pattern: "^cmo_[A-Za-z0-9_-]{32,128}$"
      responses:
        "200":
          description: Sanitized overlay state from stream-session audit logs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OverlayState"
        "404":
          description: Overlay token not found
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    adminApiKey:
      type: apiKey
      in: header
      name: x-admin-api-key
  schemas:
...
    OverlayConfigUpdate:
      type: object
      additionalProperties: false
      properties:
        enabled:
          type: boolean
        theme:
          type: string
          enum:
            - control_room
            - transparent_minimal
            - high_contrast
        activeSessionId:
          type:
            - string
            - "null"
        showModerationActions:
          type: boolean
        showRuntimeStatus:
          type: boolean
        showViewerStats:
          type: boolean
        showRecentChat:
          description: Off by default so public overlay URLs do not expose chat text unless the creator opts in.
          type: boolean
    OverlayConfig:
      type: object
      required:
        - profileId
        - configured
        - enabled
        - theme
        - activeSessionId
        - showModerationActions
        - showRuntimeStatus
        - showViewerStats
        - showRecentChat
        - tokenPreview
        - createdAt
        - updatedAt
      properties:
        profileId:
          type: string
        configured:
          type: boolean
        enabled:
          type: boolean
        theme:
          type: string
          enum:
            - control_room
            - transparent_minimal
            - high_contrast
        activeSessionId:
          type:
            - string
            - "null"
        showModerationActions:
          type: boolean
        showRuntimeStatus:
          type: boolean
        showViewerStats:
          type: boolean
        showRecentChat:
          type: boolean
        tokenPreview:
          type:
            - string
            - "null"
        createdAt:
          type:
            - string
            - "null"
          format: date-time
        updatedAt:
          type:
            - string
            - "null"
          format: date-time
        allowed:
          type: boolean
        requiredPlan:
          type:
            - string
            - "null"
    OverlayConfigMutation:
      allOf:
        - $ref: "#/components/schemas/OverlayConfig"
        - type: object
          required:
            - publicUrl
            - publicPath
            - tokenRotated
          properties:
            publicUrl:
              type:
                - string
                - "null"
              format: uri
            publicPath:
              type:
                - string
                - "null"
            tokenRotated:
              type: boolean
    OverlayState:
      type: object
      required:
        - generatedAt
        - enabled
        - status
        - overlay
        - session
        - metrics
        - recentActions
        - recentRuntimeEvents
        - recentMessages
      properties:
        generatedAt:
          type: string
          format: date-time
        enabled:
          type: boolean
        status:
          type: string
          enum:
            - disabled
            - waiting_for_stream
            - live
            - ended
        overlay:
          type: object
        session:
          type:
            - object
            - "null"
        metrics:
          type: object
          required:
            - messages
            - uniqueChatters
            - moderationActions
            - destructiveActions
            - spamAttempts
            - commandsSent
            - timersSent
            - reconnectEvents
          properties:
            messages:
              type: integer
            uniqueChatters:
              type: integer
            moderationActions:
              type: integer
            destructiveActions:
              type: integer
            spamAttempts:
              type: integer
            commandsSent:
              type: integer
            timersSent:
              type: integer
            reconnectEvents:
              type: integer
        recentActions:
          type: array
          items:
            type: object
        recentRuntimeEvents:
          type: array
          items:
            type: object
        recentMessages:
          type: array
          items:
            type: object
...